filler: switch to a pure JS auto filler
Glucose is fairly slow, and in its current implementation does
not support scored wordlists.  On a test puzzle of mine, glucose
fails to find a solution within the 30 second timeout.

I have written a purpose-built, fast javascript solver which finds
a solution in 7 seconds on the same puzzle with the same wordlist.
Moreover, this implementation supports asynchronous updates
(giving visual feedback of the in-progress solve), and scored
wordlists, although neither of those features are yet used in this
change.

Switch Phil to use the javascript solver.